+++ samples/DotNetCoreSample/Program.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Serilog;
 using WeihanLi.Common;
 using WeihanLi.Common.Event;
 using WeihanLi.Common.Helpers;
-using WeihanLi.Common.Logging.Serilog;
+using WeihanLi.Common.Logging;
+using WeihanLi.Common.Logging.Log4Net;
+using WeihanLi.Extensions;
 
 // ReSharper disable once LocalizableElement
 namespace DotNetCoreSample
@@ -17,14 +18,14 @@ namespace DotNetCoreSample
             Console.WriteLine("----------DotNetCoreSample----------");
 
             // LogHelper.AddLogProvider(new Log4NetLogHelperProvider());
-            // LogHelper.LogFactory.AddLog4Net();
-            LogHelper.LogFactory.AddSerilog(loggerConfig => loggerConfig.WriteTo.Console());
+            LogHelper.LogFactory.AddLog4Net();
+            // LogHelper.LogFactory.AddSerilog(loggerConfig => loggerConfig.WriteTo.Console());
 
             // var dataLogger = LogHelper.GetLogger(typeof(DataExtension));
             // DataExtension.CommandLogAction = msg => dataLogger.Debug(msg);
 
-            var serviceCollection = new ServiceCollection();
-            serviceCollection.AddScoped<IFly, MonkeyKing>();
+            var services = new ServiceCollection();
+            services.AddScoped<IFly, MonkeyKing>();
             IConfiguration configuration = new ConfigurationBuilder()
                 .AddJsonFile("appsettings.json")
                 .Build();
@@ -33,15 +34,18 @@ namespace DotNetCoreSample
             //var number = configuration.GetAppSetting<int>("Number");
             //Console.WriteLine($"City:{city}, Number:{number}");
 
-            serviceCollection.AddSingleton(configuration);
+            services.AddSingleton(configuration);
 
-            serviceCollection.AddSingleton<CounterEventHandler1>();
-            serviceCollection.AddSingleton<CounterEventHandler2>();
+            services.AddSingleton<IEventStore, EventStoreInMemory>();
+            services.AddSingleton<IEventBus, EventBus>();
 
-            serviceCollection.AddSingleton<IEventStore, EventStoreInMemory>();
-            serviceCollection.AddSingleton<IEventBus, EventBus>();
+            services.AddSingleton(DelegateEventHandler.FromAction<CounterEvent>(@event =>
+                LogHelper.GetLogger(typeof(DelegateEventHandler<CounterEvent>))
+                    .Info($"Event Info: {@event.ToJson()}")
+                )
+            );
 
-            DependencyResolver.SetDependencyResolver(serviceCollection);
+            DependencyResolver.SetDependencyResolver(services);
 
             //DependencyInjectionTest.Test();