+++ src/WeihanLi.Common/Event/DelegateEventHandler.cs
@@ -1,11 +1,26 @@
 ﻿using System;
 using System.Threading.Tasks;
+using WeihanLi.Extensions;
 
 namespace WeihanLi.Common.Event
 {
-    public class DelegateEventHandler<TEvent> : IEventHandler<TEvent> where TEvent : EventBase
+    public static class DelegateEventHandler
+    {
+        public static DelegateEventHandler<TEvent> FromAction<TEvent>(Action<TEvent> action) where TEvent : IEventBase => new DelegateEventHandler<TEvent>(action);
+
+        public static DelegateEventHandler<TEvent> FromFunc<TEvent>(Func<TEvent, Task> func) where TEvent : IEventBase => new DelegateEventHandler<TEvent>(func);
+    }
+
+    public class DelegateEventHandler<TEvent> : IEventHandler<TEvent> where TEvent : IEventBase
     {
         private readonly Func<TEvent, Task> _func;
+
+        public DelegateEventHandler(Action<TEvent> action)
+        {
+            if (null == action) throw new ArgumentNullException(nameof(action));
+
+            _func = action.WrapTask();
+        }
 
         public DelegateEventHandler(Func<TEvent, Task> func)
         {