+++ src/WeihanLi.Common/Event/EventBase.cs
@@ -3,7 +3,20 @@ using Newtonsoft.Json;
 
 namespace WeihanLi.Common.Event
 {
-    public abstract class EventBase
+    public interface IEventBase
+    {
+        /// <summary>
+        /// Event publish time
+        /// </summary>
+        DateTimeOffset EventAt { get; }
+
+        /// <summary>
+        /// eventId
+        /// </summary>
+        string EventId { get; }
+    }
+
+    public abstract class EventBase : IEventBase
     {
         [JsonProperty]
         public DateTimeOffset EventAt { get; private set; }
@@ -11,7 +24,7 @@ namespace WeihanLi.Common.Event
         [JsonProperty]
         public string EventId { get; private set; }
 
-        public EventBase()
+        protected EventBase()
         {
             EventId = GuidIdGenerator.Instance.NewId();
             EventAt = DateTimeOffset.UtcNow;