+++ src/WeihanLi.Common/Event/IEventBus.cs
@@ -3,27 +3,27 @@
     public interface IEventBus
     {
         /// <summary>
-        /// register event handler
+        /// add event handler for event
         /// </summary>
         /// <typeparam name="TEvent">TEvent</typeparam>
         /// <typeparam name="TEventHandler">TEventHandler</typeparam>
         /// <returns>whether the operation success</returns>
-        bool Subscribe<TEvent, TEventHandler>() where TEventHandler : IEventHandler<TEvent> where TEvent : EventBase;
+        bool Subscribe<TEvent, TEventHandler>() where TEventHandler : IEventHandler<TEvent> where TEvent : IEventBase;
 
         /// <summary>
-        /// unregister event handler
+        /// remove event handler for event
         /// </summary>
         /// <typeparam name="TEvent">TEvent</typeparam>
         /// <typeparam name="TEventHandler">TEventHandler</typeparam>
         /// <returns>whether the operation success</returns>
-        bool Unsubscribe<TEvent, TEventHandler>() where TEventHandler : IEventHandler<TEvent> where TEvent : EventBase;
+        bool Unsubscribe<TEvent, TEventHandler>() where TEventHandler : IEventHandler<TEvent> where TEvent : IEventBase;
 
         /// <summary>
-        /// Publish event
+        /// publish event
         /// </summary>
         /// <typeparam name="TEvent">event type</typeparam>
         /// <param name="event">event</param>
         /// <returns>whether the operation success</returns>
-        bool Publish<TEvent>(TEvent @event) where TEvent : EventBase;
+        bool Publish<TEvent>(TEvent @event) where TEvent : IEventBase;
     }
 }