+++ src/WeihanLi.Common/Event/IEventHandler.cs
@@ -2,8 +2,12 @@
 
 namespace WeihanLi.Common.Event
 {
-    public interface IEventHandler<TEvent> where TEvent : EventBase
+    public interface IEventHandler<in TEvent> where TEvent : IEventBase
     {
+        /// <summary>
+        /// Handler event
+        /// </summary>
+        /// <param name="event">event</param>
         Task Handle(TEvent @event);
     }
 }