+++ src/WeihanLi.Common/Helpers/TaskHelper.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace WeihanLi.Common.Helpers
+{
+    public static class TaskHelper
+    {
+        public static Task<T> FromDefault<T>() => Task.FromResult(default(T));
+    }
+}